Fix warning on using char as array index

+++ arch/i386/kbd.c
@@ -57,7 +57,7 @@ void kbd_isr_main(void)
 		if (keycode < 0)
 			return;
 
-		putchar(kbd_map[keycode]);
+		putchar(kbd_map[(int)keycode]);
 	}
 
 }
